Added permissions for map, maps API key, and imported maps library

+++ Course/src/edu/westmont/course/ShowLocationActivity.java
@@ -69,8 +69,8 @@ public class ShowLocationActivity extends Activity implements LocationListener {
 
 	  @Override
 	  public void onLocationChanged(Location location) {
-	    int lat = (int) (location.getLatitude());
-	    int lng = (int) (location.getLongitude());
+	    double lat =  (location.getLatitude());
+	    double lng =  (location.getLongitude());
 	    latituteField.setText(String.valueOf(lat));
 	    longitudeField.setText(String.valueOf(lng));
 	  }
